requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
chromadb
torch
sentence-transformers
python-dotenv